Move _path to lib/path
lib is a more typical directory for internal code.
